test: ViewingRooms
